Eliminación SP

+++ FRIO.MAR.APPLICATION.CORE/Interfaces/Repositories/IRolRepository.cs
@@ -10,7 +10,7 @@ namespace FRIO.MAR.APPLICATION.CORE.Interfaces.Repositories
     public interface IRolRepository
     {
         bool ActualizaRol(SPRol rol, ref string mensaje);
-        short? AsignarVentanas(long idRol, string idPermisos, long usuarioAuditoria, ref string mensaje);
+        int? AsignarVentanas(long idRol, string idPermisos, long usuarioAuditoria, ref string mensaje);
         bool RegistrarRol(SPRol rol, ref string mensaje);
         List<Rol> ObtenerCodigoRol();
         List<Rol> GetRoles();
